crew-ai/agent-example-001/pyproject.toml
[project]
name = "agent-example-001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.157.0",
    "flask>=3.1.1",
    "openai>=1.99.5",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
